Workaround for  https://github.com/jupyter/notebook/issues/699
This is a quick workaround for https://github.com/jupyter/notebook/issues/699, Auto-Scrolling/Jumping to Previous Location Link which renders the toc extension quite unusable https://github.com/ipython-contrib/IPython-notebook-extensions/issues/424. The issue only affects Chrome/Chromium. The workaround consists in issuing a $.ajax(), which if I understand well, resets the queue of ajax requests. The suggestion comes from @jhamrick in the discussion referenced above. This is called, after a small delay, after each click on a link in the toc window.  It seems to work (at least for me); hope it will also for others.