Fix limit output missing braces
Due to the currently mismatching braces, the limit output extension
will usually display an error message indicating it is limiting output,
although it is not (but not indicating such in console.log).

Fixed this bug and also added ability to display the current and maximum
output lengths in the error message to make this type of error easier
to debug in future.

Change-Id: Ib806f011fc4919d1a1b6279c9e37e4564d12bbed