Added Shape as a different column
Simply Added it to the html with a check that the kernel is python.
I don't know if R has shapes or anything of that sort so I add the column
only if the kernel lang is python.
If an object is not an nd.array it will print "None" which can be fixed by
making the function _getshapeof(x) return "" instead of None.