toc2: allow building toc based on arbitrary html tags
A dom_search_pattern is introduced in order to build
the ToC by searching for arbitrary html tags.

For instance, searching for captions instead of headings
allow changing the table of content in a table of figures.

Typical uses cases is for building table of recommendations,
of todo items, of issues, of references...